backend/requirements.txt
fastapi
uvicorn
elasticsearch
openai
python-dotenv
tiktoken
sentence-transformers
huggingface_hub